[#62] SCP upload with `recursive: true` should…
…notify user of an error when the destination directory does not exist.

Fix handling of warnings on SCP upload. Before, after the warning
message was collected, the upload state did not proceed to the next
upload item causing the transfer to freeze or time out. Now, whenever a
warning is encountered, the upload collects the message into the upload
state and proceeds with the next item on the upload stack.